+++ src/Solucao.cpp
@@ -3,7 +3,7 @@
 void impSol(Solucao &sol)
 {
     cout << "[ ";
-    for (int i = 0; i < sol.size(); i++)
+    for (long unsigned int i = 0; i < sol.size(); i++)
     {
         cout << sol.at(i) << " , ";
     }
